making size_t as return type of getIntersectionNum

+++ include/geos/geomgraph/Edge.h
@@ -196,16 +196,16 @@ public:
 	 * Adds EdgeIntersections for one or both
 	 * intersections found for a segment of an edge to the edge intersection list.
 	 */
-	virtual void addIntersections(algorithm::LineIntersector *li, int segmentIndex,
-		int geomIndex);
+	virtual void addIntersections(algorithm::LineIntersector *li, size_t segmentIndex,
+		size_t geomIndex);
 
 	/// Add an EdgeIntersection for intersection intIndex.
 	//
 	/// An intersection that falls exactly on a vertex of the edge is normalized
 	/// to use the higher of the two possible segmentIndexes
 	///
-	virtual void addIntersection(algorithm::LineIntersector *li, int segmentIndex,
-		int geomIndex, int intIndex);
+	virtual void addIntersection(algorithm::LineIntersector *li, size_t segmentIndex,
+		size_t geomIndex, size_t intIndex);
 
 	/// Update the IM with the contribution for this component.
 	//
